#40: Required Maven version
Added minimum required Maven version to the plugin POM in order to avoid
usage with unsupported Maven installations.
Everything below 3.2.1 is not supported since there is an version
conflict with the CDI API. It is not clear why this conflict rises up
but it must be caused by some Plexus components providing CDI stuff.
Then the classloader contains different types for CDI's TypeLiteral