#68: Plugin need property resolution for inactive profiles
Implemented a property resolver that is used to check version properties
of dependencies, plugins and plugin-dependencies of inactive profiles.
This implementation can now handle call snapshot occurrences in all
profiles of the projects, except the ones that are not added as
dependencies but part of plugin configurations, such as artifactItems of
the unpack goal of the dependencies-plugin.

Still need to implement and fix tests before building the release.